Refactoring av vurderingArtikkel16Vedtak.js, splitter opp child components i egne filer. La til publiserStegdata kall i vedKlikk metoden for å oppdatere lovvalgsperiode før fatting av vedtak
