+++ src/sider/eu_eøs/saksbehandling/stegMap/artikkel16_vedtak.js
@@ -1,6 +1,6 @@
 import Steg from "../../../../felleskomponenter/stegvelger/stegMotor/steg";
 import { FANE_STATUS, STEG } from "../../../../felleskomponenter/stegvelger";
-import VurderingArtikkel16Vedtak from "../../stegKomponenter/vurderingArtikkel16Vedtak";
+import VurderingArtikkel16Vedtak from "../../stegKomponenter/vurderingArtikkel16Vedtak/vurderingArtikkel16Vedtak";
 
 class Artikkel16Vedtak extends Steg {
   constructor(propsLight, stegPosisjon) {
@@ -19,6 +19,7 @@ class Artikkel16Vedtak extends Steg {
       lagreOgFatteVedtak: this._propsLight.tilgjengeligeHandlers.lagreOgFatteVedtak,
       tilbake: propsLight.tilgjengeligeHandlers.tilbake,
       kontrollerFerdigbehandling: this._propsLight.tilgjengeligeHandlers.kontrollerFerdigbehandling,
+      publiserStegdata: this._propsLight.tilgjengeligeHandlers.publiserStegdata,
     };
     this._status = FANE_STATUS.OK;
   }
